+++ gnubgmodule.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: gnubgmodule.c,v 1.188 2017/02/19 15:36:16 plm Exp $
+ * $Id: gnubgmodule.c,v 1.189 2017/08/01 21:52:14 plm Exp $
  */
 
 #include "config.h"
@@ -787,7 +787,7 @@ PythonHint_Callback (procrecorddata *pr)
     PyObject *list = (PyObject *)pr->pvUserData;
     PyObject *hintdict = NULL, *ctxdict = NULL, *details = NULL;
 
-    int index = (int)(long)pr->avOutputData[PROCREC_HINT_ARGOUT_INDEX];
+    int index = (int)(ptrdiff_t)pr->avOutputData[PROCREC_HINT_ARGOUT_INDEX];
     const matchstate * pms = pr->avOutputData[PROCREC_HINT_ARGOUT_MATCHSTATE];
     const movelist  * pml = pr->avOutputData[PROCREC_HINT_ARGOUT_MOVELIST];
     const evalsetup * pes = &pml->amMoves[index].esMove;
@@ -879,7 +879,7 @@ PythonHint(PyObject * UNUSED(self), PyObject * args)
         prochint.pvUserData = PyList_New(0);
         prochint.pfProcessRecord = PythonHint_Callback;
         prochint.avInputData[PROCREC_HINT_ARGIN_SHOWPROGRESS] = (void *)(long)0;
-        prochint.avInputData[PROCREC_HINT_ARGIN_MAXMOVES] = (void *)(long)nMaxMoves;
+        prochint.avInputData[PROCREC_HINT_ARGIN_MAXMOVES] = (void *)(ptrdiff_t)nMaxMoves;
         hint_move (szNumber, FALSE, (void *)&prochint);
         if (fInterrupt) {
             ResetInterrupt();
